refactor: move drag-n-drop to shared layer
