--- src/entities/drag-n-drop/index.ts
@@ -1,3 +1,0 @@
-import Draggable from "./Draggable";
-
-export default Draggable;
