+++ src/shared/lib/index.ts
@@ -1,0 +1,3 @@
+import Draggable from "./drag-n-drop/Draggable";
+
+export default Draggable;
